requirements.txt
asyncio
aiohttp
obsws-python
google-auth-oauthlib
google-api-python-client
google-api-python-client-stubs 
dataclasses-json
rich
sqlalchemy